Split portal.py and migrate more stuff to mautrix-0.4
